Move OfstedExtensions to own class
